+++ src/components/safe-messages/MsgType/styles.module.css
@@ -1,4 +1,4 @@
-.txType {
+.type {
   display: flex;
   align-items: center;
   gap: var(--space-1);
